fixed texts

+++ test/CDB_LogBinsTest.sql
@@ -1,5 +1,5 @@
 WITH data AS (
-    SELECT array_agg(x) x FROM generate_series(1,100) x
+    SELECT array_agg(x::numeric) x FROM generate_series(1,100) x
         WHERE x % 5 != 0 AND x % 7 != 0
     )
-SELECT unnest(CDB_LogBins(x::numeric, 7,''::text)) FROM data
+SELECT unnest(CDB_LogBins(x, 7,''::text)) FROM data
