+++ test/CDB_StdDevBinsTest.sql
@@ -1,5 +1,5 @@
 WITH data AS (
-    SELECT array_agg(x) x FROM generate_series(1,100) x
+    SELECT array_agg(x::numeric) x FROM generate_series(1,100) x
         WHERE x % 5 != 0 AND x % 7 != 0
     )
-SELECT unnest(CDB_StdDevBins(x::numeric)) FROM data;
+SELECT unnest(CDB_StdDevBins(x)) FROM data;
